+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Add Rank field to a Validator
+- Add validator rank endpoint to get the rank of a validator address in a specific board
 
 ## [0.1.0] - 2021-04-19